refactor(config): express S3 region relaxation via ValidateIf
Revert the IsAWSRegion decorator to its original, generic form and drop
the cross-field argument. Instead, skip the AWS-format check directly at
the call site when a custom STORAGE_S3_ENDPOINT is set, using the same
ValidateIf idiom every other config variable uses. Native AWS S3 keeps
its strict region validation; S3-compatible providers (Scaleway fr-par)
are no longer rejected.
